Feat(Favorite Articles)Add Favorite articles Functionality
- Add a model for favorite Articles
- Add a serializer to serialize the data
- Add a view for the favorite Articles
- Add a helper to assist in Functionality
- Add tests for the feature to make sure it behaves as expected.

[Delivers #162633507]
