+++ .gitignore
@@ -94,18 +94,8 @@ env/
 
 # SQLite3
 db.sqlite3
-<<<<<<< HEAD
-=======
-<<<<<<< HEAD
-=======
->>>>>>> f8271b7... Feat(Favorite Articles)Add Favorite articles Functionality
 authors/apps/authentication/social_auth/__pycache__/*.pyc
 authors/apps/authentication/tests/__pycache__/*.pyc
 
 #VS code
 *.vscode
-
-<<<<<<< HEAD
-=======
->>>>>>> ceee466... feat Enable facebook and google social auth login
->>>>>>> f8271b7... Feat(Favorite Articles)Add Favorite articles Functionality
