+++ requirements.txt
@@ -18,14 +18,10 @@ django-cors-middleware==1.3.1
 django-currentuser==0.2.3
 django-extensions==2.1.4
 django-nose==1.4.6
-<<<<<<< HEAD
-django==2.0.10
-=======
 django-oauth-toolkit==1.2.0
 django-rest-auth==0.9.3
 django-rest-framework-social-oauth2==1.1.0
 django==2.1.5
->>>>>>> f8271b7... Feat(Favorite Articles)Add Favorite articles Functionality
 djangorestframework==3.9.0
 docopt==0.6.2
 facebook-sdk==3.1.0
@@ -40,11 +36,7 @@ isort==4.3.4
 lazy-object-proxy==1.3.1
 mccabe==0.6.1
 nose==1.3.7
-<<<<<<< HEAD
 oauthlib==3.0.0           # via requests-oauthlib
-=======
-oauthlib==3.0.0
->>>>>>> f8271b7... Feat(Favorite Articles)Add Favorite articles Functionality
 pillow==5.4.1
 psycopg2-binary==2.7.6.1
 psycopg2==2.7.6.1
@@ -55,14 +47,9 @@ pyjwt==1.7.1
 pylint==2.2.2
 python-decouple==3.1
 python-twitter==3.5
-<<<<<<< HEAD
-pytz==2018.9
-requests-oauthlib==1.0.0  # via google-auth-oauthlib, python-twitter
-=======
 python3-openid==3.1.0     # via social-auth-core
 requests-oauthlib==1.0.0  # via google-auth-oauthlib, python-twitter, social-auth-core
 pytz==2018.9
->>>>>>> f8271b7... Feat(Favorite Articles)Add Favorite articles Functionality
 requests==2.21.0
 rsa==4.0                  # via google-auth
 six==1.12.0
